Support basic add,delete and edit for expense entries
Initial commit. Walking skeleton that can add and edit expense entry with an amount. The date will be set to today automatically. Category is not implemented yet. Clicking on an existing entry will open the edit page. Date picker is not implemented, edit will change date to a hard-coded date.